requirements.txt
twikit>=2.3.0
yt-dlp>=2024.1.1
requests>=2.31.0
python-dotenv>=1.0.0
